+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
-- [command_toolbox.tphere] feature: a new `useful` and `convenient` command named `/tphere <player>` to teleport `others` to `you`. (#360)
+- [command_toolbox.tppos] feature: new `/tppos here <target players>` command. (To teleport `others` to `you`.)